crypto: use safe promises for await/return within async functions
Both awaiting and returning a Promise within the body of an async
function are vulnerable to prototype pollution, which can result in
interception of intermediate values and internal state.

Returning a SafePromise from an _async_ function does not expose the
SafePromise to userland; it is wrapped safely in a new Promise.